PBI: Make the on-heap array implementation support base indicies outside the signed int range
This solves a problem where big imports would use chunked array data
structures, run low on memory, and try to allocate on-heap chunks at very large
base offsets.

+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/cache/HeapByteArray.java
@@ -32,7 +32,7 @@ public class HeapByteArray extends HeapNumberArray<ByteArray> implements ByteArr
     private final byte[] defaultValue;
     private final boolean defaultValueIsUniform;
 
-    public HeapByteArray( int length, byte[] defaultValue, int base )
+    public HeapByteArray( int length, byte[] defaultValue, long base )
     {
         super( defaultValue.length, base );
         this.length = length;
